# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API provided by DTI project",
        "title": "Domain Threat Intelligence API",
        "contact": {
            "name": "Yaroslav Lysak",
            "url": "https://t.me/Qvineox"
        },
        "version": "0.0.4"
    },
    "paths": {
...
        "/blacklists/stats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns data containing overall amount of blacklisted entities",
                "tags": [
                    "Blacklists"
                ],
                "summary": "Returns amount of blacklisted entities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routing.BlacklistedStatistics"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
...
        "/system/dynamic": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Gets info about current dynamic application config",
                "tags": [
                    "Configuration"
                ],
                "summary": "View application dynamic config",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/system/dynamic/naumen": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates dynamic Naumen Service Desk configuration",
                "tags": [
                    "Configuration"
                ],
                "summary": "Update dynamic Naumen Service Desk configuration",
                "parameters": [
                    {
                        "description": "dynamic naumen configuration",
                        "name": "naumenConfig",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routing.naumenConfigUpdateParams"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/system/dynamic/naumen/blacklists": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates dynamic Naumen Service Desk service configuration",
                "tags": [
                    "Configuration"
                ],
                "summary": "Update dynamic Naumen Service Desk service configuration",
                "parameters": [
                    {
                        "description": "dynamic naumen service configuration",
                        "name": "naumenConfig",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routing.naumenBlacklistServiceConfigUpdateParams"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/system/dynamic/reset": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Resets all dynamic configuration variables",
                "tags": [
                    "Configuration"
                ],
                "summary": "Return all dynamic configuration variables to default",
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/system/dynamic/smtp": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates dynamic SMTP configuration",
                "tags": [
                    "Configuration"
                ],
                "summary": "Update dynamic SMTP configuration",
                "parameters": [
                    {
                        "description": "dynamic SMTP configuration",
                        "name": "smtpConfig",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routing.smtpConfigUpdateParams"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "error.APIError": {
            "type": "object",
            "properties": {
                "ErrorCode": {
                    "$ref": "#/definitions/error.CodeError"
                },
                "ErrorMessage": {
                    "type": "string"
                },
                "ErrorModule": {
                    "type": "string"
                },
                "StatusCode": {
                    "type": "integer"
                }
            }
        },
        "error.CodeError": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14,
                15
            ],
            "x-enum-varnames": [
                "EncodingErrorCode",
                "DecodingErrorCode",
                "IncorrectParamsErrorCode",
                "InsufficientParamsErrorCode",
                "DatabaseErrorCode",
                "DatabaseEntityNotFoundCode",
                "FileDecodingErrorCode",
                "FileReadingErrorCode",
                "FileExtensionNotSupportedErrorCode",
                "NotImplementedErrorCode",
                "FileProcessingErrorCode",
                "InternalUnidentifiedErrorCode",
                "AuthFailedErrorCode",
                "AuthPermissionInsufficientErrorCode",
                "JobQueueErrorCode"
            ]
        },
...
        "routing.BlacklistedStatistics": {
            "type": "object",
            "properties": {
                "CreatedByDate": {
                    "$ref": "#/definitions/routing.HostsByDate"
                },
                "DiscoveredByDate": {
                    "$ref": "#/definitions/routing.HostsByDate"
                },
                "LastEval": {
                    "type": "string"
                },
                "TotalDomains": {
                    "type": "integer"
                },
                "TotalEmails": {
                    "type": "integer"
                },
                "TotalIPs": {
                    "type": "integer"
                },
                "TotalURLs": {
                    "type": "integer"
                }
            }
        },
        "routing.HostsByDate": {
            "type": "object",
            "properties": {
                "Dates": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "Domains": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "Emails": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "IPs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "URLs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "routing.naumenBlacklistServiceConfigUpdateParams": {
            "type": "object",
            "required": [
                "AgreementID",
                "CallType",
                "HostTypes",
                "SLM"
            ],
            "properties": {
                "AgreementID": {
                    "type": "integer"
                },
                "CallType": {
                    "type": "string"
                },
                "HostTypes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "SLM": {
                    "type": "integer"
                }
            }
        },
        "routing.naumenConfigUpdateParams": {
            "type": "object",
            "required": [
                "ClientGroupID",
                "ClientID",
                "ClientKey",
                "URL"
            ],
            "properties": {
                "ClientGroupID": {
                    "type": "integer"
                },
                "ClientID": {
                    "type": "integer"
                },
                "ClientKey": {
                    "type": "string"
                },
                "Enabled": {
                    "type": "boolean"
                },
                "URL": {
                    "type": "string"
                }
            }
        },
...
        "routing.smtpConfigUpdateParams": {
            "type": "object",
            "required": [
                "From",
                "Host",
                "Port"
            ],
            "properties": {
                "Enabled": {
                    "type": "boolean"
                },
                "From": {
                    "type": "string"
                },
                "Host": {
                    "type": "string"
                },
                "Password": {
                    "type": "string"
                },
                "Port": {
                    "type": "integer"
                },
                "SSL": {
                    "type": "boolean"
                },
                "UseAuth": {
                    "type": "boolean"
                },
                "User": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-api-Key",
            "in": "header"
        }
    }
}